Move TransformSpecToElements to paymentsheet module.
