+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/KlarnaHeaderStaticTextSpec.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.ui.core.elements
 
+import androidx.annotation.RestrictTo
 import com.stripe.android.uicore.elements.FormElement
 import com.stripe.android.uicore.elements.IdentifierSpec
 import kotlinx.parcelize.Parcelize
@@ -11,7 +12,8 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 @Parcelize
-internal data class KlarnaHeaderStaticTextSpec(
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+data class KlarnaHeaderStaticTextSpec(
     @SerialName("api_path")
     override val apiPath: IdentifierSpec = IdentifierSpec.Generic("klarna_header_text")
 ) : FormItemSpec() {
